refactor(helpers): split monolith into subpackage, add migration CLI
Refactor 1,970-line helpers.py into a focused helpers/ subpackage with
12 domain-specific modules. Consolidate 4 duplicate copies of stat data
into a single canonical source in _stat_data.py, eliminating ~800 lines
of redundancy including the 489-line inline STATS_NAME_MAP in
StatCalc/calculator.py. All existing import paths preserved via
backward-compatible re-export shim.

Add swgoh-migrate CLI tool (python -m swgoh_comlink.migrate) for
scanning user codebases to identify deprecated imports, API changes,
and migration steps needed when upgrading from v1.x.

Rewrite legacy test_get_enums and test_get_unit_stats from
requests/mock.patch to pytest-httpx, fixing 5 test failures.
