+++ src/swgoh_comlink/migrate/__init__.py
@@ -1,0 +1,6 @@
+# coding=utf-8
+"""Migration checker for swgoh_comlink upgrades."""
+
+from .__main__ import main
+
+__all__ = ["main"]
