+++ tests/test_get_enums.py
@@ -1,75 +1,76 @@
-from unittest import TestCase, main, mock
+"""Tests for get_enums() endpoint."""
 
-import requests
+import json
+
+import httpx
+import pytest
+from pytest_httpx import HTTPXMock
 
 from swgoh_comlink import SwgohComlink
 from swgoh_comlink.exceptions import SwgohComlinkException
 
 
-class TestGetEnums(TestCase):
-    @mock.patch("requests.request")
-    def test_get_enums(self, mock_request):
-        """Test that get_enums() makes a GET request via _request() and returns parsed JSON."""
-        mock_response = mock.Mock()
-        mock_response.content = b'{"CombatType": {"1": "CHARACTER", "2": "SHIP"}}'
-        mock_request.return_value = mock_response
+def test_get_enums(httpx_mock: HTTPXMock):
+    """Test that get_enums() makes a GET request and returns parsed JSON."""
+    httpx_mock.add_response(json={"CombatType": {"1": "CHARACTER", "2": "SHIP"}})
 
-        comlink = SwgohComlink()
-        result = comlink.get_enums()
+    client = SwgohComlink(url="http://localhost:3000")
+    result = client.get_enums()
 
-        mock_request.assert_called_once_with("GET", "http://localhost:3000/enums", headers={}, verify=True)
-        self.assertIn("CombatType", result)
-
-    @mock.patch("requests.request")
-    def test_get_enums_with_hmac(self, mock_request):
-        """Test that get_enums() applies HMAC authentication when keys are set."""
-        mock_response = mock.Mock()
-        mock_response.content = b'{"CombatType": {"1": "CHARACTER"}}'
-        mock_request.return_value = mock_response
-
-        comlink = SwgohComlink(access_key="test_access", secret_key="test_secret")
-        comlink.get_enums()
-
-        call_args = mock_request.call_args
-        self.assertEqual(call_args[0][0], "GET")
-        self.assertEqual(call_args[0][1], "http://localhost:3000/enums")
-        headers = call_args[1]["headers"]
-        self.assertIn("X-Date", headers)
-        self.assertIn("Authorization", headers)
-        self.assertTrue(headers["Authorization"].startswith("HMAC-SHA256 Credential=test_access,Signature="))
-
-    @mock.patch("requests.request")
-    def test_get_enums_respects_verify_ssl(self, mock_request):
-        """Test that get_enums() passes verify_ssl=False when configured."""
-        mock_response = mock.Mock()
-        mock_response.content = b"{}"
-        mock_request.return_value = mock_response
-
-        comlink = SwgohComlink(verify_ssl=False)
-        comlink.get_enums()
-
-        mock_request.assert_called_once_with("GET", "http://localhost:3000/enums", headers={}, verify=False)
-
-    @mock.patch("requests.request", side_effect=requests.ConnectionError("Connection refused"))
-    def test_get_enums_connection_error(self, mock_request):
-        """Test that get_enums() wraps connection errors in SwgohComlinkException."""
-        comlink = SwgohComlink()
-        with self.assertRaises(SwgohComlinkException):
-            comlink.get_enums()
-
-    @mock.patch("requests.request")
-    def test_get_enums_no_json_body_for_get(self, mock_request):
-        """Test that GET requests do not include a json body."""
-        mock_response = mock.Mock()
-        mock_response.content = b"{}"
-        mock_request.return_value = mock_response
-
-        comlink = SwgohComlink()
-        comlink.get_enums()
-
-        call_kwargs = mock_request.call_args[1]
-        self.assertNotIn("json", call_kwargs)
+    request = httpx_mock.get_request()
+    assert request.method == "GET"
+    assert str(request.url) == "http://localhost:3000/enums"
+    assert "CombatType" in result
 
 
-if __name__ == "__main__":
-    main()
+def test_get_enums_with_hmac(httpx_mock: HTTPXMock):
+    """Test that get_enums() applies HMAC authentication when keys are set."""
+    httpx_mock.add_response(json={"CombatType": {"1": "CHARACTER"}})
+
+    client = SwgohComlink(
+        url="http://localhost:3000",
+        access_key="test_access",
+        secret_key="test_secret",
+    )
+    client.get_enums()
+
+    request = httpx_mock.get_request()
+    assert request.method == "GET"
+    assert str(request.url) == "http://localhost:3000/enums"
+    assert "x-date" in request.headers
+    assert "authorization" in request.headers
+    assert request.headers["authorization"].startswith(
+        "HMAC-SHA256 Credential=test_access,Signature="
+    )
+
+
+def test_get_enums_respects_verify_ssl(httpx_mock: HTTPXMock):
+    """Test that get_enums() works with verify_ssl=False."""
+    httpx_mock.add_response(json={})
+
+    client = SwgohComlink(url="http://localhost:3000", verify_ssl=False)
+    result = client.get_enums()
+
+    request = httpx_mock.get_request()
+    assert request.method == "GET"
+    assert result == {}
+
+
+def test_get_enums_connection_error(httpx_mock: HTTPXMock):
+    """Test that get_enums() wraps connection errors in SwgohComlinkException."""
+    httpx_mock.add_exception(httpx.ConnectError("Connection refused"))
+
+    client = SwgohComlink(url="http://localhost:3000")
+    with pytest.raises(SwgohComlinkException):
+        client.get_enums()
+
+
+def test_get_enums_no_json_body_for_get(httpx_mock: HTTPXMock):
+    """Test that GET requests do not include a JSON body."""
+    httpx_mock.add_response(json={})
+
+    client = SwgohComlink(url="http://localhost:3000")
+    client.get_enums()
+
+    request = httpx_mock.get_request()
+    assert request.content == b""
